refactor(jobs): isolate claim dispatch flows
